+++ eslint.config.mjs
@@ -36,7 +36,7 @@ export default [
   {
     name: 'project-specific',
     rules: {
-      complexity: ['error', { max: 15 }]
+      complexity: ['error', { max: 20 }]
     }
   },
   {